+++ src/store.js
@@ -1,11 +1,11 @@
-const TOGGLE = 'action/toggle';
-const LOAD_TODOS = 'action/load_todos';
-const SWAP_ITEMS = 'action/swap';
-const ADD_TODO = 'action/add_todo';
-const EDIT_TODO = 'action/edit_description';
-const DELETE_TODO = 'action/delete_todo';
+export const TOGGLE = 'action/toggle';
+export const LOAD_TODOS = 'action/load_todos';
+export const SWAP_TODOS = 'action/swap';
+export const ADD_TODO = 'action/add_todo';
+export const EDIT_TODO = 'action/edit_description';
+export const DELETE_TODO = 'action/delete_todo';
 
-function createStore() {
+export default function createStore() {
   let state = [];
   const subscribers = [];
 
@@ -24,7 +24,7 @@ function createStore() {
         state = action.items;
         break;
       }
-      case SWAP_ITEMS: {
+      case SWAP_TODOS: {
         // Get values
         const src = state[action.source];
         const dest = state[action.dest];
@@ -72,64 +72,3 @@ function createStore() {
     dispatch,
   };
 }
-
-class TodoStore {
-  constructor() {
-    this.store = createStore();
-  }
-
-  get todos() {
-    return this.store.getState();
-  }
-
-  toggleTodo(index) {
-    this.store.dispatch({
-      type: TOGGLE,
-      index,
-    });
-  }
-
-  loadTodos(items) {
-    this.store.dispatch({
-      type: LOAD_TODOS,
-      items,
-    });
-  }
-
-  swapTodo(source, dest) {
-    this.store.dispatch({
-      type: SWAP_ITEMS,
-      source,
-      dest,
-    });
-  }
-
-  addTodo(text) {
-    this.store.dispatch({
-      type: ADD_TODO,
-      text,
-    });
-  }
-
-  editTodo(index, text) {
-    this.store.dispatch({
-      type: EDIT_TODO,
-      index,
-      text,
-    });
-  }
-
-  deleteTodo(index) {
-    this.store.dispatch({
-      type: DELETE_TODO,
-      index,
-    });
-  }
-
-  onUpdate(callback) {
-    this.store.subscribe(callback);
-  }
-}
-
-const store = new TodoStore();
-export default store;
